Layer infra created

+++ src/main/java/br/com/ilima/apibusiness/data/dto/ClientePedidoFormDTO.java
@@ -1,10 +1,11 @@
 package br.com.ilima.apibusiness.data.dto;
 
+import br.com.ilima.apibusiness.domain.entity.Cliente;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
 public record ClientePedidoFormDTO(
     @NotNull
     @Positive
-    Integer codigo
+    Long codigo
 ) { }
